+++ package.json
@@ -22,7 +22,7 @@
     "cookie": "^0.6.0",
     "cors": "^2.8.5",
     "moment": "^2.30.1",
-    "pinia": "^2.2.2",
+    "pinia": "^2.2.4",
     "rollup": "^4.22.4",
     "typescript": "^5.6.2",
     "vue": "^3.4.29",
