+++ src/router/index.js
@@ -38,61 +38,80 @@ import SavingListView from '@/views/financialProducts/SavingListView.vue';
 import FundListView from '@/views/financialProducts/FundListView.vue';
 import StockList from '@/views/stock/StockList.vue';
 import StockSearch from '@/views/stock/StockSearch.vue';
-import recentView from "@/views/recentView/recentView.vue";
-import portfolioData from "@/views/portfolio/PortfolioData.vue";
-import portfolioEx from "@/views/portfolio/PortfolioEx.vue";
-import cartEx from "@/views/cart/CartEx.vue";
-
+import recentView from '@/views/recentView/recentView.vue';
+import cartEx from '@/views/cart/CartEx.vue';
 
 const routes = [
-    { path: '/', name: 'Home', component: Home },
-    { path: '/find', name: 'Find', component: FindId },
-    { path: '/login', name: 'Login', component: Login },
-    { path: '/signup', name: 'SignUp', component: SignUp },
-    { path: '/mypage', name: 'MyPage', component: MyPage },
-    { path: '/mypageEdit', name: 'MyPageEdit', component: MyPageEdit },
-    { path: '/changePW', name: 'ChangePW', component: ChangePW },
-    { path: '/deleteAccount', name: 'DeleteAccount', component: DeleteAccount },
-    {
-        path: '/financial-products/:productType/:pageNumber?',
-        name: 'FinancialProducts',
-        component: FinancialProducts,
-    },
-    {
-        path: '/list/:id',
-        name: 'ProductDetail',
-        component: ProductDesc,
-    },
-    { path: '/cart', name: 'Cart', component: Cart },
-    { path: '/recently-viewed', name: 'RecentlyViewed', component: RecentlyViewed },
-    { path: '/influencer', name: 'Influencer', component: Influencer },
-    { path: '/news', name: 'News', component: News },
-    { path: '/youtube', name: 'Youtube', component: Youtube },
-    { path: '/investment-test', name: 'InvestmentTest', component: InvestmentTest },
-    { path: '/investment-test-end', name: 'InvestmentTestEnd', component: InvestmentTestEnd },
-    { path: '/investment-test-start', name: 'InvestmentTestStart', component: InvestmentTestStart },
-    { path: '/my-investment-analyze', name: 'MyInvestmentAnalyze', component: MyInvestmentAnalyze },
-    { path: '/portfolio', name: 'Portfolio', component: Portfolio },
-    { path: '/portfolio/:id', name: 'Portfolio', component: Portfolio },
-    { path: '/make-portfolio', name: 'MakePortfolio', component: MakePortfolio },
-    { path: '/my-portfolio', name: 'MyPortfolio', component: MyPortfolio },
-    { path: '/product-comparison', name: 'ProductComparison', component: ProductComparison },
-    { path: '/bonds', name: 'BondListView', component: BondListView },
-    { path: '/deposits', name: 'DepositListView', component: DepositListView },
-    { path: '/savings', name: 'SavingListView', component: SavingListView },
-    { path: '/funds', name: 'FundListView', component: FundListView },
-    { path: '/stocks', name: 'StockListView', component: StockList },
-    { path: '/stock-search', name: 'StockSearch', component: StockSearch },
-    { path: '/recent-view', name: 'RecentView', component: recentView },
-    { path: '/portfolio-data', name: 'PortfolioData', component: portfolioData },
-    { path: '/portfolio-ex', name: 'PortfolioEx', component: portfolioEx },
-    { path: '/cart-ex', name: 'CartEx', component: cartEx },
+  { path: '/', name: 'Home', component: Home },
+  { path: '/find', name: 'Find', component: FindId },
+  { path: '/login', name: 'Login', component: Login },
+  { path: '/signup', name: 'SignUp', component: SignUp },
+  { path: '/mypage', name: 'MyPage', component: MyPage },
+  { path: '/mypageEdit', name: 'MyPageEdit', component: MyPageEdit },
+  { path: '/changePW', name: 'ChangePW', component: ChangePW },
+  { path: '/deleteAccount', name: 'DeleteAccount', component: DeleteAccount },
+  {
+    path: '/financial-products/:productType/:pageNumber?',
+    name: 'FinancialProducts',
+    component: FinancialProducts,
+  },
+  {
+    path: '/list/:id',
+    name: 'ProductDetail',
+    component: ProductDesc,
+  },
+  { path: '/cart', name: 'Cart', component: Cart },
+  {
+    path: '/recently-viewed',
+    name: 'RecentlyViewed',
+    component: RecentlyViewed,
+  },
+  { path: '/influencer', name: 'Influencer', component: Influencer },
+  { path: '/news', name: 'News', component: News },
+  { path: '/youtube', name: 'Youtube', component: Youtube },
+  {
+    path: '/investment-test',
+    name: 'InvestmentTest',
+    component: InvestmentTest,
+  },
+  {
+    path: '/investment-test-end',
+    name: 'InvestmentTestEnd',
+    component: InvestmentTestEnd,
+  },
+  {
+    path: '/investment-test-start',
+    name: 'InvestmentTestStart',
+    component: InvestmentTestStart,
+  },
+  {
+    path: '/my-investment-analyze',
+    name: 'MyInvestmentAnalyze',
+    component: MyInvestmentAnalyze,
+  },
+  { path: '/portfolio', name: 'Portfolio', component: Portfolio },
+  { path: '/portfolio/:id', name: 'Portfolio', component: Portfolio },
+  { path: '/make-portfolio', name: 'MakePortfolio', component: MakePortfolio },
+  { path: '/my-portfolio', name: 'MyPortfolio', component: MyPortfolio },
+  {
+    path: '/product-comparison',
+    name: 'ProductComparison',
+    component: ProductComparison,
+  },
+  { path: '/bonds', name: 'BondListView', component: BondListView },
+  { path: '/deposits', name: 'DepositListView', component: DepositListView },
+  { path: '/savings', name: 'SavingListView', component: SavingListView },
+  { path: '/funds', name: 'FundListView', component: FundListView },
+  { path: '/stocks', name: 'StockListView', component: StockList },
+  { path: '/stock-search', name: 'StockSearch', component: StockSearch },
+  { path: '/recent-view', name: 'RecentView', component: recentView },
+
+  { path: '/cart-ex', name: 'CartEx', component: cartEx },
 ];
 
 const router = createRouter({
-    history: createWebHistory(),
-    routes,
+  history: createWebHistory(),
+  routes,
 });
 
 export default router;
-
