+++ src/store/modules/portfolio.js
@@ -1,81 +1,74 @@
-import { fetchPortfolioList, getPortfolioDetail, postPortfolio, deletePortfolio } from '@/api/portfolioApi.js';
+import { defineStore } from 'pinia';
+import { ref } from 'vue';
+import portfolioApi from '@/api/portfolioApi.js';
 
-const portfolioModule = {
-    namespaced: true,
-    state: () => ({
-        portfolioList: [],
-        portfolioListLoaded: false,
-        portfolioDetail: null,
-        newPortfolio: null,
-        newPortfolioItems: [],
-    }),
+export const usePortfolioStore = defineStore('portfolio', () => {
+  // 상태 정의
+  const portfolioList = ref([]); // 포트폴리오 목록
+  const portfolioListLoaded = ref(false); // 포트폴리오 목록이 로드되었는지 여부
+  const portfolioDetail = ref(null); // 개별 포트폴리오 상세 정보
+  const newPortfolio = ref(null); // 새로 생성된 포트폴리오
+  const newPortfolioItems = ref([]); // 새로운 포트폴리오 항목
 
-    actions: {
-        async fetchPortfolioList({ commit, state }) {
-            if(!state.portfolioListLoaded) {
-                try {
-                    const data = await fetchPortfolioList();
-                    commit('setPortfolioList', data);
-                } catch (error) {
-                    console.error('Error fetching portfolio list:', error);
-                }
-            }
-        },
+  // 포트폴리오 목록을 가져오는 함수
+  const fetchPortfolioListAction = async () => {
+    if (!portfolioListLoaded.value) {
+      try {
+        const data = await portfolioApi.fetchPortfolioList(); // API 호출
+        portfolioList.value = data; // 목록 저장
+        portfolioListLoaded.value = true; // 로드 완료 표시
+      } catch (error) {
+        console.error('Error fetching portfolio list:', error);
+      }
+    }
+  };
 
-        async getPortfolioDetail({ commit }, portfolioId) {
-            try {
-                const data = await getPortfolioDetail(portfolioId);
-                commit('setPortfolioDetail', data);
-            } catch (error) {
-                console.error('Error fetching portfolio detail:', error);
-            }
-        },
+  // 개별 포트폴리오 상세 정보를 가져오는 함수
+  const getPortfolioDetailAction = async (portfolioId) => {
+    try {
+      const data = await portfolioApi.getPortfolioDetail(portfolioId); // 상세 정보 가져오기
+      portfolioDetail.value = data; // 상태에 저장
+    } catch (error) {
+      console.error('Error fetching portfolio detail:', error);
+    }
+  };
 
-        async postPortfolio({ commit }, portfolio) {
-            try {
-                const data = await postPortfolio(portfolio);
-                commit('setNewPortfolio', data);
-            } catch (error) {
-                console.error('Error posting portfolio:', error);
-            }
-        },
+  // 새로운 포트폴리오를 생성하는 함수
+  const postPortfolioAction = async (portfolio) => {
+    try {
+      const data = await portfolioApi.postPortfolio(portfolio); // 새 포트폴리오 생성
+      newPortfolio.value = data; // 상태에 저장
+      portfolioList.value.push(data); // 포트폴리오 목록에 새 항목 추가
+    } catch (error) {
+      console.error('Error posting portfolio:', error);
+    }
+  };
 
-        async deletePortfolio({ commit }, portfolioId) {
-            try {
-                const data = await deletePortfolio(portfolioId);
-                commit('removePortfolio', portfolioId);
-            } catch (error) {
-                console.error('Error deleting portfolio:', error);
-            }
-        },
-    },
+  // 포트폴리오를 삭제하는 함수
+  const deletePortfolioAction = async (portfolioId) => {
+    console.log(`Attempting to delete portfolio: ${portfolioId}`);
+    try {
+      await portfolioApi.deletePortfolio(portfolioId); // 삭제 요청
+      portfolioList.value = portfolioList.value.filter(
+        (portfolio) => portfolio.id !== portfolioId
+      );
+      console.log(`Portfolio ${portfolioId} deleted successfully.`);
+    } catch (error) {
+      console.error(`Error deleting portfolio: ${portfolioId}`, error);
+      throw error; // 오류를 상위로 던져서 처리
+    }
+  };
 
-    getters: {
-        portfolioList: (state) => state.portfolioList,
-        portfolioDetail: (state) => state.portfolioDetail,
-        newPortfolio: (state) => state.newPortfolio,
-        isPortfolioListLoaded: (state) => state.portfolioListLoaded,
-    },
-
-    mutations: {
-        setPortfolioList(state, portfolioList) {
-            state.portfolioList = portfolioList;
-            state.portfolioListLoaded = true;
-        },
-
-        setPortfolioDetail(state, portfolioDetail) {
-            state.portfolioDetail = portfolioDetail;
-        },
-
-        setNewPortfolio(state, newPortfolio) {
-            state.newPortfolio = newPortfolio;
-            state.portfolioList.push(newPortfolio); // 새로운 포트폴리오 리스트에 추가
-        },
-
-        removePortfolio(state, portfolioId) {
-            state.portfolioList = state.portfolioList.filter(
-                (portfolio) => portfolio.id !== portfolioId
-            );
-        },
-    },
-};
+  return {
+    portfolioList,
+    portfolioListLoaded,
+    portfolioDetail,
+    newPortfolio,
+    newPortfolioItems,
+    fetchPortfolioListAction,
+    getPortfolioDetailAction,
+    postPortfolioAction,
+    deletePortfolioAction,
+  };
+});
+export default usePortfolioStore; // 추가된 default export
